#1 Made the navigation system with react navigation
o Made two screens. One for the Map and one for the Calendar
o Inserted them into the navigation that we use in App.js
o Imported the correct icons so that it works with Expo. Got some errors here when importing icons from react-native-elements. So changed to some Expo friendly
o Inserted the icons to the navigation settings, and added some settings so we can se which screen in selected down in the navigation bar as well.