#7 Added the calendar to CalendarScreen
o Used react-native-calendars to add a new Calendar component
o I think we should make a new screen AgendaScreen, that we show each time we press a date in the Calendar. This is used in examples in the github repo for react-native-calendars
o Added the theme settings, so we all can see how we easily can change them.
o Changed the arrow icons in the calendar to our Expo icons